Refactor: extract liveobjects steps classes
